debugging part 34

+++ backend/db/migrations/20241216171214-create-spot.js
@@ -42,11 +42,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(10,7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       lng: {
-        type: Sequelize.DECIMAL(10,7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       name: {
@@ -58,7 +58,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.DECIMAL(7,2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
